services/placepicker/requirements.txt
Flask==0.10.1
python-google-places==1.2.0
